% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCanonicalPathways.R
\name{plotCanonicalPathways}
\alias{plotCanonicalPathways}
\alias{plotCanonicalPathways.data.frame}
\alias{plotCanonicalPathways.list}
\title{plotCanonicalPathways}
\usage{
plotCanonicalPathways(plot_data, ...)

\method{plotCanonicalPathways}{data.frame}(plot_data,
  number_pathways = 5, ...)

\method{plotCanonicalPathways}{list}(plot_data, number_pathways = 5, ...)
}
\arguments{
\item{plot_data}{Either one data.frame/tibble or a list of data.frames/tibbles to plot.
Accepts data from \link[IPAPlots]{prepCanonicalPathways}}

\item{...}{Additional parameters to pass}

\item{number_pathways}{Number of pathways to plot.  Default: 5}
}
\description{
plotCanonicalPathways
}
