% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegulatorZscores.R
\name{plotRegulatorZscores}
\alias{plotRegulatorZscores}
\alias{plotRegulatorZscores.data.frame}
\alias{plotRegulatorZscores.list}
\title{plotRegulatorZscores}
\usage{
plotRegulatorZscores(...)

\method{plotRegulatorZscores}{data.frame}(module_table,
  number_genes_show = 10, pos_label_force = 10, neg_label_force = 10,
  plot_these_regulators = NULL, title = NULL, ...)

\method{plotRegulatorZscores}{list}(module_table, ...)
}
\arguments{
\item{...}{Additional parameters to pass}

\item{module_table}{A data.frame or tibble with the columns "Upstream Regulator", "Activation z-score",
and "p-value of overlap".  Use \link[IPAPlots]{prepUpstreamRegulatorLists}}

\item{number_genes_show}{The number of genes with the top negative and positive z-scores to label}

\item{pos_label_force}{The replusion force for the positive z-scoring gene names, passed to geom_text_repel}

\item{neg_label_force}{The replusion force for the negative z-scoring gene names, passed to geom_text_repel}

\item{plot_these_regulators}{A list of potential regulators to use as a filter.  Only those regulators
that are in this list will be displayed (if present).}

\item{title}{Plot title}
}
\value{
a ggplot object
}
\description{
plotRegulatorZscores
}
