% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepCanonicalPathways.R
\name{prepCanonicalPathways}
\alias{prepCanonicalPathways}
\title{prepCanonicalPathways}
\usage{
prepCanonicalPathways(path, filename_extension = "txt")
}
\arguments{
\item{path}{Path to files exported from IPA}

\item{filename_extension}{Extension on output files.  Used to filter files in the directory.  Default: txt}
}
\description{
Read in data for Canonical pathways as exported by IPA and prepare for plotting
}
