% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepUpstreamRegulatorLists.R
\name{prepUpstreamRegulatorLists}
\alias{prepUpstreamRegulatorLists}
\title{prepUpstreamRegulatorLists}
\usage{
prepUpstreamRegulatorLists(path, filename_extension = "txt")
}
\arguments{
\item{path}{The path to a directory containing output files from IPA}

\item{filename_extension}{File extension used for the IPA output files.  Default: "txt"}
}
\value{
list of data.frames, each of which can be plotted using \link[IPAPlots]{plotRegulatorZscores}
}
\description{
Read in a directory populated with upstream regulator tables from IPA
and return a named list ready for making plots
}
\examples{
reg_lists <- prepUpstreamRegulatorList(path = "~/path_to_ipa_output")
}
