% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatModuleNames.R
\name{reformatModuleNames}
\alias{reformatModuleNames}
\title{reformatModuleNames}
\usage{
reformatModuleNames(urlist_names)
}
\arguments{
\item{urlist_names}{List of character vectors to rename.}
}
\value{
List of character vectors.
}
\description{
Reformats shorthand used for file names to text for titles
}
