% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSeedInfo.R
\name{removeSeedInfo}
\alias{removeSeedInfo}
\title{removeSeedInfo}
\usage{
removeSeedInfo(urlists)
}
\arguments{
\item{urlists}{}
}
\value{
list
}
\description{
This gets rid of the "(and other miRNAs w/seed GGAAUGU)"-like info after all the miRNAs
}
