% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCanonicalPlots.R
\name{saveCanonicalPlots}
\alias{saveCanonicalPlots}
\title{saveCanonicalPlots}
\usage{
saveCanonicalPlots(plot_list, output_dir = "~/", width = 8,
  height = 8, units = "in", ...)
}
\arguments{
\item{plot_list}{List of plots generated by \code[IPAPlots]{plotCanonicalPathways}}

\item{output_dir}{Directory to save the plots in}

\item{width}{Width of the saved plots.  Default: 8.}

\item{height}{Height of the saved plots.  Default: 8.}

\item{units}{Units used for the width and height.  Default: "in"}

\item{...}{Additional parameters to pass to ggsave}
}
\description{
Save plots as generated by \code[IPAPlots]{plotCanonicalPathways} to pngs.
}
